backend/requirements.txt
# Core
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-gis>=1.1.0
black==24.2.0
psycopg2-binary==2.9.9
GDAL>=3.0.0

# Development
flake8==7.0.0
black==24.2.0
factory-boy>=3.0.0

# API Documentation & Development Tools
django-cors-headers==4.3.1
drf-spectacular==0.27.1